requirements.txt
fastapi[standard]==0.115.8
pydantic-settings
pymongo
